@@ -5,7 +5,7 @@
 	<modelVersion>4.0.0</modelVersion>
 	<groupId>io.goobi.viewer.module</groupId>
 	<artifactId>module-skeleton</artifactId>
-	<version>21.11-SNAPSHOT</version>
+	<version>23.04-SNAPSHOT</version>
 	<packaging>jar</packaging>
 
 
@@ -15,17 +15,18 @@
 
 	<properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <jdk.version>1.8</jdk.version>
+        <jdk.version>11</jdk.version>
         
         <!-- Manifest information -->
         <maven.build.timestamp.format>yyyy-MM-dd HH:mm</maven.build.timestamp.format>
         <buildDate>${maven.build.timestamp}</buildDate>
         
-		<viewer-core.version>21.11-SNAPSHOT</viewer-core.version>
-		<dbunit.version>2.7.2</dbunit.version> 
+		<viewer-core.version>23.04-SNAPSHOT</viewer-core.version>
+		<dbunit.version>2.7.3</dbunit.version> 
 		<h2.version>1.4.200</h2.version>
-        <mockito.version>4.1.0</mockito.version>
-        <solr.version>8.11.0</solr.version>
+		<log4j.version>2.20.0</log4j.version>
+        <mockito.version>5.2.0</mockito.version>
+        <solr.version>8.11.2</solr.version>
 	</properties>
 	
 	
